split out route config parsing
