+++ CMakeLists.txt
@@ -2668,6 +2668,7 @@ add_library(grpc
   src/core/xds/grpc/xds_lb_policy_registry.cc
   src/core/xds/grpc/xds_listener.cc
   src/core/xds/grpc/xds_route_config.cc
+  src/core/xds/grpc/xds_route_config_parser.cc
   src/core/xds/grpc/xds_routing.cc
   src/core/xds/grpc/xds_transport_grpc.cc
   src/core/xds/xds_client/xds_api.cc
