+++ build_autogenerated.yaml
@@ -1240,6 +1240,7 @@ libs:
   - src/core/xds/grpc/xds_lb_policy_registry.h
   - src/core/xds/grpc/xds_listener.h
   - src/core/xds/grpc/xds_route_config.h
+  - src/core/xds/grpc/xds_route_config_parser.h
   - src/core/xds/grpc/xds_routing.h
   - src/core/xds/grpc/xds_transport_grpc.h
   - src/core/xds/xds_client/xds_api.h
@@ -2046,6 +2047,7 @@ libs:
   - src/core/xds/grpc/xds_lb_policy_registry.cc
   - src/core/xds/grpc/xds_listener.cc
   - src/core/xds/grpc/xds_route_config.cc
+  - src/core/xds/grpc/xds_route_config_parser.cc
   - src/core/xds/grpc/xds_routing.cc
   - src/core/xds/grpc/xds_transport_grpc.cc
   - src/core/xds/xds_client/xds_api.cc
