+++ gRPC-C++.podspec
@@ -1349,6 +1349,7 @@ Pod::Spec.new do |s|
                       'src/core/xds/grpc/xds_lb_policy_registry.h',
                       'src/core/xds/grpc/xds_listener.h',
                       'src/core/xds/grpc/xds_route_config.h',
+                      'src/core/xds/grpc/xds_route_config_parser.h',
                       'src/core/xds/grpc/xds_routing.h',
                       'src/core/xds/grpc/xds_transport_grpc.h',
                       'src/core/xds/xds_client/xds_api.h',
@@ -2628,6 +2629,7 @@ Pod::Spec.new do |s|
                               'src/core/xds/grpc/xds_lb_policy_registry.h',
                               'src/core/xds/grpc/xds_listener.h',
                               'src/core/xds/grpc/xds_route_config.h',
+                              'src/core/xds/grpc/xds_route_config_parser.h',
                               'src/core/xds/grpc/xds_routing.h',
                               'src/core/xds/grpc/xds_transport_grpc.h',
                               'src/core/xds/xds_client/xds_api.h',
