+++ gRPC-Core.podspec
@@ -2115,6 +2115,8 @@ Pod::Spec.new do |s|
                       'src/core/xds/grpc/xds_listener.h',
                       'src/core/xds/grpc/xds_route_config.cc',
                       'src/core/xds/grpc/xds_route_config.h',
+                      'src/core/xds/grpc/xds_route_config_parser.cc',
+                      'src/core/xds/grpc/xds_route_config_parser.h',
                       'src/core/xds/grpc/xds_routing.cc',
                       'src/core/xds/grpc/xds_routing.h',
                       'src/core/xds/grpc/xds_transport_grpc.cc',
@@ -3403,6 +3405,7 @@ Pod::Spec.new do |s|
                               'src/core/xds/grpc/xds_lb_policy_registry.h',
                               'src/core/xds/grpc/xds_listener.h',
                               'src/core/xds/grpc/xds_route_config.h',
+                              'src/core/xds/grpc/xds_route_config_parser.h',
                               'src/core/xds/grpc/xds_routing.h',
                               'src/core/xds/grpc/xds_transport_grpc.h',
                               'src/core/xds/xds_client/xds_api.h',
