+++ src/core/BUILD
@@ -5276,6 +5276,33 @@ grpc_cc_library(
 )
 
 grpc_cc_library(
+    name = "xds_route_config",
+    srcs = [
+        "xds/grpc/xds_route_config.cc",
+    ],
+    hdrs = [
+        "xds/grpc/xds_route_config.h",
+    ],
+    external_deps = [
+        "absl/strings",
+        "absl/strings:str_format",
+        "absl/types:optional",
+        "absl/types:variant",
+        "re2",
+    ],
+    language = "c++",
+    tags = ["nofixdeps"],
+    deps = [
+        "grpc_matchers",
+        "match",
+        "time",
+        "xds_http_filter",
+        "//:grpc_base",
+        "//:xds_client",
+    ],
+)
+
+grpc_cc_library(
     name = "grpc_xds_client",
     srcs = [
         "xds/grpc/xds_audit_logger_registry.cc",
@@ -5292,7 +5319,7 @@ grpc_cc_library(
         "xds/grpc/xds_http_stateful_session_filter.cc",
         "xds/grpc/xds_lb_policy_registry.cc",
         "xds/grpc/xds_listener.cc",
-        "xds/grpc/xds_route_config.cc",
+        "xds/grpc/xds_route_config_parser.cc",
         "xds/grpc/xds_routing.cc",
         "xds/grpc/xds_transport_grpc.cc",
     ],
@@ -5311,7 +5338,7 @@ grpc_cc_library(
         "xds/grpc/xds_http_stateful_session_filter.h",
         "xds/grpc/xds_lb_policy_registry.h",
         "xds/grpc/xds_listener.h",
-        "xds/grpc/xds_route_config.h",
+        "xds/grpc/xds_route_config_parser.h",
         "xds/grpc/xds_routing.h",
         "xds/grpc/xds_transport_grpc.h",
     ],
@@ -5450,6 +5477,7 @@ grpc_cc_library(
         "xds_credentials",
         "xds_file_watcher_certificate_provider_factory",
         "xds_http_filter",
+        "xds_route_config",
         "xds_type_upb",
         "xds_type_upbdefs",
         "//:channel",
@@ -5543,6 +5571,7 @@ grpc_cc_library(
         "xds_common_types",
         "xds_credentials",
         "xds_http_filter",
+        "xds_route_config",
         "//:api_trace",
         "//:config",
         "//:debug_location",
@@ -6764,6 +6793,7 @@ grpc_cc_library(
         "grpc_xds_client",
         "match",
         "ref_counted",
+        "xds_route_config",
         "//:config",
         "//:gpr",
         "//:grpc_resolver",
@@ -6813,6 +6843,7 @@ grpc_cc_library(
         "time",
         "xds_dependency_manager",
         "xds_http_filter",
+        "xds_route_config",
         "xxhash_inline",
         "//:channel_arg_names",
         "//:config",
