+++ tools/doxygen/Doxyfile.c++.internal
@@ -3007,6 +3007,8 @@ src/core/xds/grpc/xds_listener.cc \
 src/core/xds/grpc/xds_listener.h \
 src/core/xds/grpc/xds_route_config.cc \
 src/core/xds/grpc/xds_route_config.h \
+src/core/xds/grpc/xds_route_config_parser.cc \
+src/core/xds/grpc/xds_route_config_parser.h \
 src/core/xds/grpc/xds_routing.cc \
 src/core/xds/grpc/xds_routing.h \
 src/core/xds/grpc/xds_transport_grpc.cc \
